+++ app/Statistics/Repository/ServerRepository.php
@@ -31,7 +31,7 @@ class ServerRepository implements ServerRepositoryInterface
     public function serverDate(): string
     {
         // TODO: Duplicates BrowserRepository::browserDate
-        return FunctionsDate::timestampToGedcomDate(WT_TIMESTAMP)->display();
+        return FunctionsDate::timestampToGedcomDate((int) WT_TIMESTAMP)->display();
     }
 
     /**
